requirements.txt
asgiref==3.2.10
bcrypt==3.2.0
cbpro==1.1.4
cffi==1.14.5
cryptography==3.4.7
Django==3.1
djangorestframework==3.12.4
Fabric3==1.14.post1
paramiko==2.7.2
pycparser==2.20
pymongo==3.5.1
PyNaCl==1.4.0
pytz==2021.1
requests==2.13.0
selenium==3.141.0
six==1.10.0
sortedcontainers==2.3.0
sqlparse==0.4.1
urllib3==1.26.4
websocket-client==0.40.0
websockets==9.0.1
